project-management-api/requirements.txt
# Web Framework & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database & ORM (with greenlet for async)
sqlalchemy==2.0.25
greenlet==3.0.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Serialization
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic[email]==2.6.3
email-validator==2.1.0

# Request Handling
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.1

# Environment Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.26.0
